Bug/WhatsNextAnimationTOOOBig
Couple of bug fixes to make What's Next section flawless.
