+++ yashadvert/src/pages/WhatsNext/WhatsNext.js
@@ -1,7 +1,52 @@
+// Author: Geon Gayles
 import React from 'react';
-
-import './style.css';
 import Slider from './slider';
+import './style.css';
+
+import deltaDental from '../../assets/images/deltaDentalC.gif';
+import angloAmerican from '../../assets/images/angloAmericanC.gif';
+import caterpillar from '../../assets/images/caterpillarC.gif';
+import johnDeere from '../../assets/images/johnDeereC.gif';
+import redLobster from '../../assets/images/redLobsterC.gif';
+import polaris from '../../assets/images/polarisC.gif';
+import johnson from '../../assets/images/jandjC.gif';
+
+const imageArray = [
+    {
+        src: deltaDental,
+        alt: "Delta Dental"
+    },
+
+    {
+        src: redLobster,
+        alt: "Red Lobster"
+    },
+
+    {
+        src: angloAmerican,
+        alt: "Anglo American"
+    },
+
+    {
+        src: caterpillar,
+        alt: "Caterpillar"
+    },
+
+    {
+        src: johnDeere,
+        alt: "John Deere"
+    },
+
+    {
+        src: polaris,
+        alt: "Polaris"
+    },
+
+    {
+        src: johnson,
+        alt: "Johnson & Johnson"
+    },
+]
 
 const WhatsNext = (props, context) => {
 
@@ -10,7 +55,7 @@ const WhatsNext = (props, context) => {
 
             <span className="title"><p>Whats Next?</p></span>
 
-            <div class="contain">
+            <div className="contain">
 
                 <h1 className="titleC">Clients</h1>
 
@@ -18,9 +63,9 @@ const WhatsNext = (props, context) => {
                     YASH Technologies has a client base of 200+ that we are constantly working with to improve and advance the technological world.
                 </p>
 
-                <div class="row">
-                    <div class="row__inner">
-                        <Slider/>
+                <div className="row">
+                    <div className="row__inner">
+                        {imageArray.map((img) => <Slider key={img.alt} src={img.src} alt={img.alt} />)}
                     </div>
                 </div>
             </div>
