+++ yashadvert/src/pages/WhatsNext/slider.js
@@ -1,64 +1,16 @@
-import React from 'react';
-import deltaDental from '../../assets/images/deltaDentalC.gif';
-import angloAmerican from '../../assets/images/angloAmericanC.gif';
-import caterpillar from '../../assets/images/caterpillarC.gif';
-import johnDeere from '../../assets/images/johnDeereC.gif';
-import redLobster from '../../assets/images/redLobsterC.gif';
-import polaris from '../../assets/images/polarisC.gif';
-import johnson from '../../assets/images/jandjC.gif';
-
-const imageArray = [
-    {
-        src: deltaDental,
-        alt: "Delta Dental"
-    },
-
-    {
-        src: redLobster,
-        alt: "Red Lobster"
-    },
-
-    {
-        src: angloAmerican,
-        alt: "Anglo American"
-    },
-
-    {
-        src: caterpillar,
-        alt: "Caterpillar"
-    },
-
-    {
-        src: johnDeere,
-        alt: "John Deere"
-    },
-
-    {
-        src: polaris,
-        alt: "Polaris"
-    },
-
-    {
-        src: johnson,
-        alt: "Johnson & Johnson"
-    },
-]
+// Author: Geon Gayles
+import React from 'react'
 
 function Slider({src, alt}) {
   return (
     
-      <div class="tile">
-      <div class="tile__media">
-        {imageArray.map((img) => {
-            return(
-            <img class="tile__img" key={img.alt} src={img.src} alt={img.alt}  />
-            )
-        })
-        }
-
+      <div className="tile">
+      <div className="tile__media">
+        <img className="tile__img" src={src} alt={alt}  />
+      </div>
+      <div className="tile__details">
         
       </div>
-      <div class="tile__details">  </div>
     </div>
     
   )
