+++ yashadvert/src/pages/WhatsNext/style.css
@@ -1,6 +1,6 @@
 /* Custom CSS For WhatsNext Section */
 
-/* Author:  */
+/* Author: Geon Gayles */
 
 
 .whats_next_content {
@@ -13,7 +13,6 @@
 
 .titleC {
     color: #002e48;
-    
 }
 
 .contain {
@@ -33,19 +32,19 @@
   padding-bottom: 10px;
 
   -webkit-animation-name: move;
-  -webkit-animation-duration: 55s;
+  -webkit-animation-duration: 48s;
   -webkit-animation-iteration-count: infinite;
   -webkit-animation-direction: right;
   -webkit-animation-timing-function: linear;
 
   -moz-animation-name: move;
-  -moz-animation-duration: 30s;
+  -moz-animation-duration: 48s;
   -moz-animation-iteration-count: infinite;
   -moz-animation-direction: right;
   -moz-animation-timing-function: linear;
 
   animation-name: move;
-  animation-duration: 55s;
+  animation-duration: 48s;
   animation-iteration-count: infinite;
   animation-direction: right;
   animation-timing-function: linear;
@@ -87,7 +86,6 @@
 .tile__details:before {
   left: 0;
   width: 100%;
-
   z-index: 2;
 }
 
@@ -107,7 +105,6 @@
 }
 .tile:hover ~ .tile {
   transform: translate3d(125px, 0, 0);
-
 }
 
 @-webkit-keyframes move {
